web/api/requirements.txt
openai==1.11.1
python-dotenv==1.0.0
supabase==2.0.3
httpx==0.24.1
pydantic==2.4.2 